fix: 500 when send /login request twice
In the RefreshToken entity class, the token field is marked as a unique key, which is achieved by the @Column(nullable = false, unique = true) annotation, while the RefreshToken entity also contains a one-to-one relationship with the TodoUser. So it's required to delete the existing refresh token before adding a new one.